requirements.txt
streamlit
pandas
pymongo
wordcloud
requests
beautifulsoup4
bs4
matplotlib
nltk
subprocess
openpyxl
uvicorn
typing
fastapi
pydantic
python-dotenv
transformers
sentencepiece
torch
torchvision
torchaudio
matplotlib
google.generativeai
pymupdf
